refactor: rename package to `com.github.catppuccin.jetbrains_icons`
