+++ .gitignore
@@ -50,4 +50,4 @@ node_modules
 
 ### Auto-generated ###
 src/main/resources/icons
-src/main/kotlin/com/github/catppuccin/icons/Icons.kt
+src/main/kotlin/com/github/catppuccin/jetbrains_icons/Icons.kt
