+++ src/main/kotlin/com/github/catppuccin/jetbrains_icons/settings/PluginSettingsState.kt
@@ -1,8 +1,7 @@
-package com.github.catppuccin.icons.settings
+package com.github.catppuccin.jetbrains_icons.settings
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
-import com.intellij.openapi.components.ServiceManager
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import com.intellij.util.xmlb.XmlSerializerUtil
